What works well: The buttons work The detector works The lock pin works The leds fade in and out As anticipated the logic module is not yet flawless, though the lack of the debouncing is the cause..
What does not go well:
After being locked the signal does not resume itss previous state
The debounce module still does not work if more than 1 object is debounced. Recommended action is to rewrite the entire module
The signal drive-on-sight works but the blinking seems to be rather sporadic from time to time
The outgoing frequency seems not to be what it supposed to be, the math should be checked
The servo had to be commented out because itss library interfered with the led fade pwm pin
After the signal is unlocked, it does not assume it's previous state.

What is not yet tested:
the fall time is not yet tested because of the failing debouncing action
The red button, because I only had 2 buttons, it propably works though.

Course of action:
The debounce module needs urgent fixing or rewritten entirely, the program heavily relies on this to work
The first logic tests show minor bugs. All signals must continue to work such as the detector or the buttons during a locked position
after the signal is unlocked the new state must be set accodingly. I recommend to create a special unlocked state or rename the current newState