Most elements of the SAS seem to work.
The buttons are read in and processed correctly including drive on sight
The fall time delay also works very well

The one thing what does not yet work is resuming the old state after being unlocked.

The servo needs a new library. I recommend to use the micros timers for it.

The fading leds work well.

The incomming frequency must also be tested, but the SAS does not recognize a frequency of 0.
recommended is to ditch the ISR and replace it by a poll based function. Perhaps use millis as well, the signal also need to be flank responsve